添加了后端版本/backend/backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
marshmallow==3.19.0
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==20.1.0
pytest==7.3.1
Werkzeug==2.2.3
